STORM-2556 Break down AutoCreds implementations into two kinds of classes
* Nimbus plugin: implements INimbusCredentialPlugin, ICredentialsRenewer
* Worker & Topology submitter: implements IAutoCredentials
* Add 'Hadoop' to abstract classes which depends on Hadoop
* also changed field name for storm-core interfaces to make them clear
  * whether it's cluster conf or topology conf
