+++ external/storm-autocreds/src/main/java/org/apache/storm/hdfs/security/AutoHDFS.java
@@ -18,50 +18,19 @@
 
 package org.apache.storm.hdfs.security;
 
-import org.apache.commons.math3.util.Pair;
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.FileSystem;
-import org.apache.hadoop.security.Credentials;
-import org.apache.hadoop.security.SecurityUtil;
-import org.apache.hadoop.security.UserGroupInformation;
-import org.apache.hadoop.security.token.Token;
-import org.apache.hadoop.security.token.TokenIdentifier;
-import org.apache.storm.Config;
-import org.apache.storm.common.AbstractAutoCreds;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.apache.storm.common.AbstractHadoopAutoCreds;
 
-import javax.security.auth.Subject;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.ObjectOutputStream;
-import java.net.URI;
-import java.security.PrivilegedAction;
-import java.util.Collection;
-import java.util.HashMap;
 import java.util.Map;
 
-import static org.apache.storm.hdfs.security.HdfsSecurityUtil.STORM_KEYTAB_FILE_KEY;
-import static org.apache.storm.hdfs.security.HdfsSecurityUtil.STORM_USER_NAME_KEY;
+import static org.apache.storm.hdfs.security.HdfsSecurityUtil.HDFS_CREDENTIALS;
 
 /**
- * Automatically get HDFS delegation tokens and push it to user's topology. The class
- * assumes that HDFS configuration files are in your class path.
+ * Auto credentials plugin for HDFS implementation. This class provides a way to automatically
+ * push credentials to a topology and to retrieve them in the worker.
  */
-public class AutoHDFS extends AbstractAutoCreds {
-    private static final Logger LOG = LoggerFactory.getLogger(AutoHDFS.class);
-    public static final String HDFS_CREDENTIALS = "HDFS_CREDENTIALS";
-    public static final String TOPOLOGY_HDFS_URI = "topology.hdfs.uri";
-
-    private String hdfsKeyTab;
-    private String hdfsPrincipal;
-
+public class AutoHDFS extends AbstractHadoopAutoCreds {
     @Override
-    public void doPrepare(Map<String, Object> conf) {
-        if(conf.containsKey(STORM_KEYTAB_FILE_KEY) && conf.containsKey(STORM_USER_NAME_KEY)) {
-            hdfsKeyTab = (String) conf.get(STORM_KEYTAB_FILE_KEY);
-            hdfsPrincipal = (String) conf.get(STORM_USER_NAME_KEY);
-        }
+    public void doPrepare(Map conf) {
     }
 
     @Override
@@ -70,147 +39,7 @@ public class AutoHDFS extends AbstractAutoCreds {
     }
 
     @Override
-    public void shutdown() {
-        //no op.
-    }
-
-    @Override
-    protected  byte[] getHadoopCredentials(Map<String, Object> conf, String configKey) {
-        Configuration configuration = getHadoopConfiguration(conf, configKey);
-        return getHadoopCredentials(conf, configuration);
-    }
-
-    @Override
-    protected byte[] getHadoopCredentials(Map<String, Object> conf) {
-        return getHadoopCredentials(conf, new Configuration());
-    }
-
-    private Configuration getHadoopConfiguration(Map<String, Object> topoConf, String configKey) {
-        Configuration configuration = new Configuration();
-        fillHadoopConfiguration(topoConf, configKey, configuration);
-        return configuration;
-    }
-
-    @SuppressWarnings("unchecked")
-    private byte[] getHadoopCredentials(Map<String, Object> conf, final Configuration configuration) {
-        try {
-            if(UserGroupInformation.isSecurityEnabled()) {
-                login(configuration);
-
-                final String topologySubmitterUser = (String) conf.get(Config.TOPOLOGY_SUBMITTER_PRINCIPAL);
-
-                final URI nameNodeURI = conf.containsKey(TOPOLOGY_HDFS_URI) ? new URI(conf.get(TOPOLOGY_HDFS_URI).toString())
-                        : FileSystem.getDefaultUri(configuration);
-
-                UserGroupInformation ugi = UserGroupInformation.getCurrentUser();
-
-                final UserGroupInformation proxyUser = UserGroupInformation.createProxyUser(topologySubmitterUser, ugi);
-
-                Credentials creds = (Credentials) proxyUser.doAs(new PrivilegedAction<Object>() {
-                    @Override
-                    public Object run() {
-                        try {
-                            FileSystem fileSystem = FileSystem.get(nameNodeURI, configuration);
-                            Credentials credential= proxyUser.getCredentials();
-
-                            if (configuration.get(STORM_USER_NAME_KEY) == null) {
-                                configuration.set(STORM_USER_NAME_KEY, hdfsPrincipal);
-                            }
-
-                            fileSystem.addDelegationTokens(configuration.get(STORM_USER_NAME_KEY), credential);
-                            LOG.info("Delegation tokens acquired for user {}", topologySubmitterUser);
-                            return credential;
-                        } catch (IOException e) {
-                            throw new RuntimeException(e);
-                        }
-                    }
-                });
-
-
-                ByteArrayOutputStream bao = new ByteArrayOutputStream();
-                ObjectOutputStream out = new ObjectOutputStream(bao);
-
-                creds.write(out);
-                out.flush();
-                out.close();
-
-                return bao.toByteArray();
-            } else {
-                throw new RuntimeException("Security is not enabled for HDFS");
-            }
-        } catch (Exception ex) {
-            throw new RuntimeException("Failed to get delegation tokens." , ex);
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    @SuppressWarnings("unchecked")
-    public void doRenew(Map<String, String> credentials, Map<String, Object> topologyConf) {
-        for (Pair<String, Credentials> cred : getCredentials(credentials)) {
-            try {
-                Configuration configuration = getHadoopConfiguration(topologyConf, cred.getFirst());
-                Collection<Token<? extends TokenIdentifier>> tokens = cred.getSecond().getAllTokens();
-
-                if (tokens != null && !tokens.isEmpty()) {
-                    for (Token token : tokens) {
-                        //We need to re-login some other thread might have logged into hadoop using
-                        // their credentials (e.g. AutoHBase might be also part of nimbu auto creds)
-                        login(configuration);
-                        long expiration = token.renew(configuration);
-                        LOG.info("HDFS delegation token renewed, new expiration time {}", expiration);
-                    }
-                } else {
-                    LOG.debug("No tokens found for credentials, skipping renewal.");
-                }
-            } catch (Exception e) {
-                LOG.warn("could not renew the credentials, one of the possible reason is tokens are beyond " +
-                        "renewal period so attempting to get new tokens.", e);
-                populateCredentials(credentials, topologyConf);
-            }
-        }
-    }
-
-    private void login(Configuration configuration) throws IOException {
-        if (configuration.get(STORM_KEYTAB_FILE_KEY) == null) {
-            configuration.set(STORM_KEYTAB_FILE_KEY, hdfsKeyTab);
-        }
-        if (configuration.get(STORM_USER_NAME_KEY) == null) {
-            configuration.set(STORM_USER_NAME_KEY, hdfsPrincipal);
-        }
-        SecurityUtil.login(configuration, STORM_KEYTAB_FILE_KEY, STORM_USER_NAME_KEY);
-
-        LOG.info("Logged into hdfs with principal {}", configuration.get(STORM_USER_NAME_KEY));
-    }
-
-    @Override
-    protected String getCredentialKey(String configKey) {
+    public String getCredentialKey(String configKey) {
         return HDFS_CREDENTIALS + configKey;
     }
-
-    @SuppressWarnings("unchecked")
-    public static void main(String[] args) throws Exception {
-        Map<String, Object> conf = new HashMap();
-        conf.put(Config.TOPOLOGY_SUBMITTER_PRINCIPAL, args[0]); //with realm e.g. storm@WITZEND.COM
-        conf.put(STORM_USER_NAME_KEY, args[1]); //with realm e.g. hdfs@WITZEND.COM
-        conf.put(STORM_KEYTAB_FILE_KEY, args[2]);// /etc/security/keytabs/storm.keytab
-
-        Configuration configuration = new Configuration();
-        AutoHDFS autoHDFS = new AutoHDFS();
-        autoHDFS.prepare(conf);
-
-        Map<String,String> creds  = new HashMap<String, String>();
-        autoHDFS.populateCredentials(creds, conf);
-        LOG.info("Got HDFS credentials", autoHDFS.getCredentials(creds));
-
-        Subject s = new Subject();
-        autoHDFS.populateSubject(s, creds);
-        LOG.info("Got a Subject "+ s);
-
-        autoHDFS.renew(creds, conf);
-        LOG.info("renewed credentials", autoHDFS.getCredentials(creds));
-    }
 }
-
